Replace ContainerSpec.DebugHelper with ContainerTests.DebugHelper.
